refactor(mercury): extract transaction related strut
