+++ mercury/model/resp/mercury_info_resp.go
@@ -1,5 +1,7 @@
 package resp
 
+import "github.com/nervosnetwork/ckb-sdk-go/mercury/model/common"
+
 type MercuryInfo struct {
 	MercuryVersion    string      `json:"mercury_version"`
 	CkbNodeVersion    string      `json:"ckb_node_version"`
@@ -17,7 +19,7 @@ const (
 )
 
 type Extension struct {
-	Name     string    `json:"name"`
-	Scripts  []Script  `json:"scripts"`
-	CellDeps []CellDep `json:"cell_deps"`
+	Name     string           `json:"name"`
+	Scripts  []common.Script  `json:"scripts"`
+	CellDeps []common.CellDep `json:"cell_deps"`
 }
