+++ mercury/model/resp/transcation_resp.go
@@ -2,43 +2,17 @@ package resp
 
 import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/nervosnetwork/ckb-sdk-go/mercury/model/common"
 	"github.com/nervosnetwork/ckb-sdk-go/types"
 )
 
 type transactionResp struct {
-	Version     hexutil.Uint    `json:"version"`
-	Hash        types.Hash   `json:"hash"`
-	CellDeps    []CellDep    `json:"cell_deps"`
-	HeaderDeps  []types.Hash `json:"header_deps"`
-	Inputs      []cellInput     `json:"inputs"`
-	Outputs     []cellOutput    `json:"outputs"`
-	OutputsData []hexutil.Bytes `json:"outputs_data"`
-	Witnesses   []hexutil.Bytes `json:"witnesses"`
-}
-
-type CellDep struct {
-	OutPoint outPoint      `json:"out_point"`
-	DepType  types.DepType `json:"dep_type"`
-}
-
-type outPoint struct {
-	TxHash types.Hash   `json:"tx_hash"`
-	Index  hexutil.Uint `json:"index"`
-}
-
-type cellInput struct {
-	Since          hexutil.Uint64 `json:"since"`
-	PreviousOutput outPoint       `json:"previous_output"`
-}
-
-type cellOutput struct {
-	Capacity hexutil.Uint64 `json:"capacity"`
-	Lock     *Script        `json:"lock"`
-	Type     *Script        `json:"type"`
-}
-
-type Script struct {
-	CodeHash types.Hash           `json:"code_hash"`
-	HashType types.ScriptHashType `json:"hash_type"`
-	Args     hexutil.Bytes        `json:"args"`
+	Version     hexutil.Uint        `json:"version"`
+	Hash        types.Hash          `json:"hash"`
+	CellDeps    []common.CellDep    `json:"cell_deps"`
+	HeaderDeps  []types.Hash        `json:"header_deps"`
+	Inputs      []common.CellInput  `json:"inputs"`
+	Outputs     []common.CellOutput `json:"outputs"`
+	OutputsData []hexutil.Bytes     `json:"outputs_data"`
+	Witnesses   []hexutil.Bytes     `json:"witnesses"`
 }
